+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+
+#### 6.4.0 / 2020-01-27
 * explicitly set file permissions to 644 for CSS files
 * updated for new v2 views or TEC and ECP
 * make legend superpowers work for new v2 views, thanks Gustavo!!